fix(#2296): remove blanket signal_toolkit mypy exclusion and fix type errors (#2339)
- Remove `ignore_errors = true` override for `signal_toolkit.*` in pyproject.toml;
  replace with targeted per-file overrides for the three mixin widget files
  (widget_processing, widget_plotting, widget_ui) that have 81 cross-mixin
  attr-defined errors pending a WidgetProtocol expansion.
- Remove `src/shared/python/` from the CI delta mypy exclusion grep filter in
  ci-standard.yml so changes to shared/python are now type-checked on PRs.
- Fix all 17 non-widget type errors:
  - calculus.py: use np.trapezoid directly (np.trapz removed in NumPy 2.0)
  - core.py: wrap np.mean() results in float() for fs/dt properties
  - noise.py: cast numpy scalar to float before passing to typed parameter
  - series.py: add type: ignore[arg-type/operator] for numpy ArrayLike boundary
  - io.py: annotate value_data dict, ignore np.savez kwargs overload mismatch,
    annotate processed Union type, fix process_all return type signature

+++ src/shared/python/signal_toolkit/noise.py
@@ -94,7 +94,7 @@ class NoiseGenerator:
             frequency = kwargs.get(
                 "frequency", DEFAULT_LINE_FREQUENCY_HZ
             )  # Default 60 Hz (line noise)
-            fs = 1.0 / np.mean(np.diff(t)) if len(t) > 1 else 1000.0
+            fs = float(1.0 / np.mean(np.diff(t))) if len(t) > 1 else 1000.0
             values = self._generate_periodic_noise(n, amplitude, frequency, fs)
 
         else:
